Update Dockerfile
In 2017 Binder had several large updates to it which changed the
underlying way it worked. As a result, Dockerfiles were no longe
expliticly required for most projects and the way in which Dockerfiles
needed to be formatted for custom projects changed as well.
c.f. https://mybinder.readthedocs.io/en/latest/dockerfile.html